Fix dcf's comment:
Personally I would remove the MUST language here as it's kind of
ineffectual--can a PT proxy reply with SIGNAL DORMANT-ENTER-DONE 60 seconds
later? Is the parent process going to wait and enforce the MUST? If not, it's
not really a MUST. I feel the MUST language will encourage PT proxies to lie
about their current state (I need to send SIGNAL DORMANT-ENTER-DONE because the
spec says so, even if I currently am not capable of going dormant). What tor is
really interested is not whether the PT proxy has received and understood the
SIGNAL DORMANT-ENTER message, it's the points in time when the PT proxy
actually changes its dormant state, so the spec should reflect that.